+++ TOMB4/tomb4/tomb4plus/t4plus_items.h
@@ -1,3 +1,3 @@
 #pragma once
 
-extern void T4PlusActivateItem(size_t item_id, bool anti);
+extern void T4PlusActivateItem(short item_id, bool anti);
